move items to pages
