+++ src/pages/Items/ItemsGen.jsx
@@ -2,12 +2,12 @@ import React, { useState, useEffect } from 'react';
 import { useHistory, useParams } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { Button } from '@material-ui/core';
-import { SaveAlert } from '../Common/Alerts';
 import {
     createItemAct,
     updateItemByIdAct,
 } from '../../state/items/itemActions';
 import { convertKeysCase } from '../../utils/caseConversion';
+import { SaveAlert } from '../../components/Common/Alerts';
 import ItemsTemplate from './ItemsTemplate';
 
 const ItemsGen = props => {
