+++ src/pages/Items/index.js
@@ -1,3 +1,2 @@
 export { default as ItemsDash } from './ItemsDash';
 export { default as ItemsGen } from './ItemsGen';
-export { default as ItemsTemplate } from './ItemsTemplate';
